mcp-server/requirements.txt
fastmcp>=2.0.0
PyGithub>=2.1.1
